requirements.txt
aerich==0.8.2
asyncpraw==7.8.1
colorama==0.4.6
disnake==2.10.1
levenshtein==0.27.1
tortoise-orm[asyncpg]==0.24.2
twitchAPI==4.4.0
